Separate out burden estimate endpoints into their own controller
